Offload class handling to Spatial.
Maintaining the components in the entity's class was becoming painful. I added a property to Spatial that stores the components in a string, which is updated when the component list changes. During render, we only have to do a string comparison to check for equality.

I also put the entity into the data structure, so we don't need to call Crafty in entity_render. All this has cut render time down to half of what it was when I started tonight.